Refactor AlertBox component
