+++ Source/DesignSystem/atoms/AlertBox/index.ts
@@ -1,4 +1,5 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export { AlertBox, AlertBoxProps, AlertBoxErrorMessage, AlertBoxInfoMessage } from './AlertBox';
+export { AlertBox, AlertBoxProps } from './AlertBox';
+export { AlertBoxErrorMessage, AlertBoxInfoMessage } from './helpers';
