+++ Source/DesignSystem/atoms/AlertDialog/helpers.ts
@@ -6,17 +6,16 @@ Dialogs are windows that appear in front of all content to provide immediate and
 to ask the user for a decision. Dialogs disable all other functionality until a user engages with it by either confirming,
 dismissing, or taking the required action. Dialogs are purposefully interruptive so they should be used sparingly.
 
-####Title
-- In some cases the description may be enough information for the user to make a decision and a title is not needed.
-- If a title is required use a clear question or statement with an explanation in the content area. Avoid apologies, 
+**Styling:** In some cases the description may be enough information for the user to make a decision and a title is not needed.
+If a title is required use a clear question or statement with an explanation in the content area. Avoid apologies, 
 ambiguity or questions such as "warning!" or "are you sure?".
 
-####Types of Dialogs
-- **Alert dialogs** are urgent interruptions, requiring acknowledgement or action, that inform the user about a situation. 
-    Most alerts do not need titles. They summarize a decision in a sentence or two by either asking a question or making a 
-    statement related to the action buttons.
-    Alert dialogs must always contain an action item such as "agree", "accept", or "continue" and a dismissive action such as "cancel", 
-    "disagree" or "deny".
+**Types of Dialogs:**
+- ***Alert dialogs*** are urgent interruptions, requiring acknowledgement or action, that inform the user about a situation. 
+Most alerts do not need titles. They summarize a decision in a sentence or two by either asking a question or making a 
+statement related to the action buttons.
+Alert dialogs must always contain an action item such as "agree", "accept", or "continue" and a dismissive action such as "cancel", 
+"disagree" or "deny".
 `;
 
 export const listArray = [
